DT-438 Mock the community API for auth tests
Previously if Delius was down the DeliusUserService just carried on
regardless which was fine.  However this behaviour has changed and
if Delius is down we thrown an exception.

Since introducing this change the first 2 tests in
AuthUserSpecification started failing intermittently with a strange
connection related error:

java.net.HttpRetryException: cannot retry due to server authentication,
in streaming mode

My theory is that the two are related and the exception somehow taints
the connection and if reused then it results in the error above.