🚧 PPD-201: Matching NOMIS and Auth users by email
Work in progress. There is not yet any way of searching for
users across both Auth and NOMIS based on their email address.
This is work in progress as we have need to search potentially
by partial email address and first / last names too.

Because this pulls data from two datasets, this just returns
a list (no pagination).